druid-admin Feature: add support for services in k8s cluster without a discovery service like eureka, nacos... Fix: 1. fix fastjson error when serialize large object. 2. fix an int overflow error

+++ druid-admin/src/main/java/com/alibaba/druid/admin/model/dto/DataSourceResult.java
@@ -118,7 +118,7 @@ public class DataSourceResult {
         @JSONField(name = "RemoveAbandoned")
         private boolean RemoveAbandoned;
         @JSONField(name = "ClobOpenCount")
-        private int ClobOpenCount;
+        private long ClobOpenCount;
         @JSONField(name = "BlobOpenCount")
         private int BlobOpenCount;
         @JSONField(name = "KeepAliveCheckCount")
